requirements.txt
numpy
pandas
pymysql
mysql-connector
python-dotenv
scikit-learn
#-e .